Ai_Fitness_Tracker/requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.11
numpy==1.26.4
speechrecognition==3.10.1
pyaudio==0.2.14
gTTS==2.5.1
playsound3==3.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
matplotlib==3.8.3
groq==0.4.2
requests==2.31.0
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic-settings==2.2.1
alembic==1.13.1
redis==5.0.3
asyncpg==0.29.0
aiosqlite==0.20.0
httpx==0.27.0
pytest==8.0.2
pytest-asyncio==0.23.5
email-validator==2.3.0
